refactor converting satoshi methods, add agreement activity

+++ app/src/main/java/io/multy/storage/RealmManager.java
@@ -49,7 +49,10 @@ public class RealmManager {
         if (realm == null || realm.isClosed()) {
             open();
         }
-        realm.executeTransaction(realm -> realm.deleteAll());
+
+        if (realm != null) {
+            realm.executeTransaction(realm -> realm.deleteAll());
+        }
     }
 
     private static void isRealmAvailable() {
